Реализована деактивация пользователя.

+++ public/js/userLk.js
@@ -479,10 +479,17 @@ forgetMeBttn.onclick = function(event){
     // Если пароль корректный, TODO деактивировать пользователя
     if(wasPapaProud){
 		if(pswdInput.value == userApi.curUser.password){
-			console.log("Деактивация");
-			socket.emit('leave', { username: userApi.curUser.login });
-			localStorage.clear();
-			window.location = "/";
+            console.log("Деактивация");
+            userApi.deactivateAccount(function (success) {
+                if (success) {
+                    socket.emit('leave', { username: userApi.curUser.login });
+                    localStorage.clear();
+                    window.location = "/";
+                } else {
+                    alert("Не удалось деактивировать аккаунт!");
+                }
+                
+            });
 		}
 		else{
 			makePapaProud(pswdInput.parentNode, false);
